requirements.txt
aiofiles==0.7.0
aiomysql==0.0.21
aiosqlite==0.16.1
amqp==5.0.6
appdirs==1.4.4
APScheduler==3.7.0
asgiref==3.3.4
atlastk==0.12.10
attrs==21.2.0
Automat==20.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
Brotli==1.0.9
celery==5.1.1
certifi==2018.1.18
cffi==1.14.5
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
cssselect==1.1.0
diff-match-patch==20200713
distlib==0.3.2
Django==2.0.2
django-apscheduler==0.3.0
django-cors-headers==2.1.0
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.0.0
django-rest-swagger==2.1.2
django-reversion==2.0.13
django-suit @ https://github.com/darklow/django-suit/tarball/v2
djangorestframework==3.7.7
filelock==3.0.12
Flask==2.0.1
Flask-Compress==1.10.1
Flask-SQLAlchemy==2.5.1
furl==2.1.2
gerapy==0.9.6
gevent==21.1.2
greenlet==1.1.0
httptools==0.3.0
hyperlink==21.0.0
idna==2.6
incremental==21.3.0
iso8601==0.1.16
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.1
json2html==1.3.0
kombu==5.1.0
logparser==0.8.2
lxml==4.6.3
MarkupSafe==2.0.1
multidict==5.1.0
openapi-codec==1.3.2
orderedmultidict==1.0.1
parsel==1.6.0
pbr==5.6.0
pexpect==4.8.0
prompt-toolkit==3.0.19
Protego==0.1.16
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pyee==8.1.0
pymongo==3.11.4
PyMySQL==0.9.3
pyOpenSSL==20.0.1
pypika-tortoise==0.1.1
pypinyin==0.36.0
pyppeteer==0.2.5
pyquery==1.4.3
python-crontab==2.2.8
python-dateutil==2.8.2
python-docx==0.8.6
python-scrapyd-api==2.1.2
pytz==2017.3
queuelib==1.6.1
redis==3.5.3
requests==2.18.4
sanic==21.6.2
Sanic-Auth==0.3.0
sanic-routing==0.7.1
Scrapy==1.8.0
scrapy-redis==0.7.1
scrapy-splash==0.7.2
scrapydweb @ git+https://github.com/my8100/scrapydweb.git@7a3b81dba2cba4279c9465064a693bb277ac20e9
service-identity==21.1.0
simplejson==3.17.3
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.20
sqlparse==0.4.1
stevedore==3.3.0
tablib==3.0.0
tortoise-orm==0.17.6
tqdm==4.61.1
Twisted==21.2.0
typing-extensions==3.10.0.0
tzlocal==2.1
ujson==4.1.0
uritemplate==3.0.1
urllib3==1.22
uvloop==0.16.0
uWSGI==2.0.18
vine==5.0.0
virtualenv==20.4.7
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
w3lib==1.22.0
wcwidth==0.2.5
websocket==0.2.1
websockets==9.1
Werkzeug==2.0.1
XlsxWriter==1.0.5
xlwt==1.3.0
zope.event==4.5.0
zope.interface==5.4.0
